backend/requirements.txt
fastapi==0.115.9
uvicorn==0.34.3
pydantic==2.11.5
python-dotenv==1.1.0
pymongo==4.13.0
chromadb==1.0.9
openai==1.86.0
langchain==0.3.25
langchain-openai==0.3.23
langgraph==0.4.8
langchain-community==0.3.25
langchain-core==0.3.65
